+++ CHANGELOG.md
@@ -3,4 +3,4 @@ Changelog
 
 ### 3.2.2.7
 
-- Removed redundant code. Hooray for ~1KB less space consumption!
+- Removed redundant code. Hooray for ~1KB less space consumption! by prasselpikachu