+++ CHANGELOG.md
@@ -4,3 +4,4 @@ Changelog
 ### 3.2.2.7
 
 - Fix issues with filesizes not showing correctly when downloading mods for a pack
+- Fix issues with username changes causing users to not be able to login with Mojang anymore