+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 Changelog
 ====================================
 
-### 3.2.2.7
+### 3.2.3.0
 
 - Fix issues with filesizes not showing correctly when downloading mods for a pack
 - Fix issues with username changes causing users to not be able to login with Mojang anymore